+++ src/Http/Controllers/RoleController.php
@@ -2,8 +2,6 @@
 
 namespace ErenMustafaOzdal\LaravelUserModule\Http\Controllers;
 
-use Illuminate\Http\Request;
-
 use App\Http\Requests;
 use Cartalyst\Sentinel\Laravel\Facades\Sentinel;
 use App\Role;
@@ -29,6 +27,7 @@ class RoleController extends AdminBaseController
      */
     public function index()
     {
+        $this->authorize(Role::take(1)->get()[0]);
         return view(config('laravel-user-module.views.role.index'));
     }
 
@@ -39,6 +38,7 @@ class RoleController extends AdminBaseController
      */
     public function create()
     {
+        $this->authorize(Role::take(1)->get()[0]);
         return view(config('laravel-user-module.views.role.create'));
     }
 
@@ -50,6 +50,7 @@ class RoleController extends AdminBaseController
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize(Role::take(1)->get()[0]);
         return $this->storeModel(Sentinel::getRoleRepository()->createModel(), $request, [
             'success'   => StoreSuccess::class,
             'fail'      => StoreFail::class
@@ -64,6 +65,7 @@ class RoleController extends AdminBaseController
      */
     public function show(Role $role)
     {
+        $this->authorize($role);
         return view(config('laravel-user-module.views.role.show'), compact('role'));
     }
 
@@ -75,6 +77,7 @@ class RoleController extends AdminBaseController
      */
     public function edit(Role $role)
     {
+        $this->authorize($role);
         return view(config('laravel-user-module.views.role.edit'), compact('role'));
     }
 
@@ -87,6 +90,7 @@ class RoleController extends AdminBaseController
      */
     public function update(UpdateRequest $request, Role $role)
     {
+        $this->authorize($role);
         return $this->updateModel($role,$request, [
             'success'   => UpdateSuccess::class,
             'fail'      => UpdateFail::class
@@ -101,6 +105,7 @@ class RoleController extends AdminBaseController
      */
     public function destroy(Role $role)
     {
+        $this->authorize($role);
         return $this->destroyModel($role, [
             'success'   => DestroySuccess::class,
             'fail'      => DestroyFail::class
