+++ src/Http/Controllers/UserApiController.php
@@ -36,6 +36,8 @@ class UserApiController extends AdminBaseController
      */
     public function index(Request $request)
     {
+        $this->authorize('indexApi', Sentinel::getUser());
+
         $users = User::select(['id','photo','first_name','last_name', 'is_active','created_at']);
         // if is filter action
         if ($request->has('action') && $request->input('action') === 'filter') {
@@ -45,7 +47,7 @@ class UserApiController extends AdminBaseController
         $addColumns = [
             'addUrls' => [
                 'activate'      => ['route' => 'api.user.activate', 'id' => true],
-                'not_activate'  => ['route' => 'api.user.not_activate', 'id' => true],
+                'not_activate'  => ['route' => 'api.user.notActivate', 'id' => true],
                 'edit_page'     => ['route' => 'admin.role.edit', 'id' => true]
             ],
             'status'            => function($model) { return $model->is_active; },
@@ -69,6 +71,7 @@ class UserApiController extends AdminBaseController
     public function detail($id, Request $request)
     {
         $user = User::where('id',$id)->with('roles')->select(['id','email','last_login','updated_at']);
+        $this->authorize('detailApi', $user->get()[0]);
 
         $editColumns = [
             'roles'         => function($model) { return $model->roles->implode('name', ', '); },
@@ -88,6 +91,7 @@ class UserApiController extends AdminBaseController
      */
     public function fastEdit(User $user, Request $request)
     {
+        $this->authorize('fastEditApi', $user);
         return $user;
     }
 
@@ -99,6 +103,7 @@ class UserApiController extends AdminBaseController
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('storeApi', Sentinel::getUser());
         return $this->storeModel(User::class, $request, [
             'success'           => StoreSuccess::class,
             'fail'              => StoreFail::class,
@@ -116,6 +121,7 @@ class UserApiController extends AdminBaseController
      */
     public function update(UpdateRequest $request, User $user)
     {
+        $this->authorize('updateApi', $user);
         $result = $this->updateModel($user, $request, [
             'success'   => UpdateSuccess::class,
             'fail'      => UpdateFail::class
@@ -140,14 +146,11 @@ class UserApiController extends AdminBaseController
      */
     public function destroy(User $user)
     {
-        if ($user->id != Sentinel::getUser()->id) {
-            return $this->destroyModel($user, [
-                'success'   => DestroySuccess::class,
-                'fail'      => DestroyFail::class
-            ]);
-        } else {
-            return response()->json(['result' => 'self']);
-        }
+        $this->authorize('destroyApi', $user);
+        return $this->destroyModel($user, [
+            'success'   => DestroySuccess::class,
+            'fail'      => DestroyFail::class
+        ]);
     }
 
     /**
@@ -158,6 +161,7 @@ class UserApiController extends AdminBaseController
      */
     public function activate(User $user)
     {
+        $this->authorize('activateApi', $user);
         $result = $this->activationComplete($user, [
             'activationSuccess'     => ActivateSuccess::class,
             'activationFail'        => ActivateFail::class
@@ -176,6 +180,7 @@ class UserApiController extends AdminBaseController
      */
     public function notActivate(User $user)
     {
+        $this->authorize('notActivateApi', $user);
         $result = $this->activationRemove($user, [
             'activationRemove'      => ActivateRemove::class,
             'activationFail'        => ActivateFail::class
@@ -194,6 +199,7 @@ class UserApiController extends AdminBaseController
      */
     public function group(Request $request)
     {
+        $this->authorize('groupApi', Sentinel::getUser());
         $events = [];
         switch($request->input('action')) {
             case 'activate':
@@ -224,6 +230,7 @@ class UserApiController extends AdminBaseController
      */
     public function destroyAvatar(FileRepository $file, Request $request, User $user)
     {
+        $this->authorize('destroyAvatarApi', $user);
         $file->deleteDirectory(config('laravel-user-module.user.uploads.path') . "/{$user->id}");
         $user->photo = NULL;
         if ( $user->save() ) {
@@ -241,6 +248,7 @@ class UserApiController extends AdminBaseController
      */
     public function avatarPhoto(PhotoRequest $request, User $user)
     {
+        $this->authorize('avatarPhotoApi', $user);
         return $this->updateModel($user, $request, [
             'success'   => UpdateSuccess::class,
             'fail'      => UpdateFail::class
