+++ src/Http/Controllers/UserController.php
@@ -3,6 +3,7 @@
 namespace ErenMustafaOzdal\LaravelUserModule\Http\Controllers;
 
 use App\Http\Requests;
+use Cartalyst\Sentinel\Laravel\Facades\Sentinel;
 use App\User;
 
 use ErenMustafaOzdal\LaravelModulesBase\Controllers\AdminBaseController;
@@ -31,6 +32,7 @@ class UserController extends AdminBaseController
      */
     public function index()
     {
+        $this->authorize(Sentinel::getUser());
         return view(config('laravel-user-module.views.user.index'));
     }
 
@@ -41,6 +43,7 @@ class UserController extends AdminBaseController
      */
     public function create()
     {
+        $this->authorize(Sentinel::getUser());
         return view(config('laravel-user-module.views.user.create'));
     }
 
@@ -52,6 +55,7 @@ class UserController extends AdminBaseController
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize(Sentinel::getUser());
         return $this->storeModel(User::class, $request, [
             'success'           => StoreSuccess::class,
             'fail'              => StoreFail::class,
@@ -68,6 +72,7 @@ class UserController extends AdminBaseController
      */
     public function show(User $user)
     {
+        $this->authorize($user);
         return view(config('laravel-user-module.views.user.show'), compact('user'));
     }
 
@@ -79,6 +84,7 @@ class UserController extends AdminBaseController
      */
     public function edit(User $user)
     {
+        $this->authorize($user);
         return view(config('laravel-user-module.views.user.edit'), compact('user'));
     }
 
@@ -91,6 +97,7 @@ class UserController extends AdminBaseController
      */
     public function update(UpdateRequest $request, User $user)
     {
+        $this->authorize($user);
         $result = $this->updateModel($user, $request,  [
             'success'   => UpdateSuccess::class,
             'fail'      => UpdateFail::class
@@ -116,6 +123,7 @@ class UserController extends AdminBaseController
      */
     public function changePassword(PasswordRequest $request, User $user)
     {
+        $this->authorize($user);
         return $this->updateModel($user,$request,  [
             'success'   => UpdateSuccess::class,
             'fail'      => UpdateFail::class
@@ -131,6 +139,7 @@ class UserController extends AdminBaseController
      */
     public function permission(PermissionRequest $request, User $user)
     {
+        $this->authorize($user);
         return $this->updateModel($user,$request,  [
             'success'   => UpdateSuccess::class,
             'fail'      => UpdateFail::class
@@ -145,6 +154,7 @@ class UserController extends AdminBaseController
      */
     public function destroy(User $user)
     {
+        $this->authorize($user);
         return $this->destroyModel($user, [
             'success'   => DestroySuccess::class,
             'fail'      => DestroyFail::class
